SL-824: Restore Linden patches to BugsplatMac upload-archive.sh.
At some point I managed to overwrite our patched version of upload-archive.sh
in BugsplatMac.framework with the plain vanilla vendor script, at which point
Mac viewer builds stopped being able to upload symbols to BugSplat because
those patches are essential to the way we invoke upload-archive.sh.

Failure to post symbol files to BugSplat is not immediately obvious because
the failure doesn't propagate up through the layers of script calls.

I was able to recover the patched script by looking for the last viewer build
that DID successfully post Mac symbols, checking the bugsplat autobuild
package tarball URL for that build and extracting its upload-archive.sh.
Naturally that had to be merged with the current upload-archive.sh to preserve
vendor changes since that earlier autobuild package was built.

Added a warning to README.txt.